refactor: move routes to utils and update imports
