+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Link } from 'react-router-dom';
 import './App.css';
-import AppRoutes from './routes';
+import AppRoutes from './utils/routes';
 
 function App() {
   return (
