+++ src/utils/routes.tsx
@@ -1,5 +1,5 @@
 import { Routes, Route } from 'react-router-dom';
-import Home from './components/Home';
+import Home from '../pages/Home';
 
 const AppRoutes = () => {
   return (
